requirements.txt
databases==0.9.0
fastapi==0.110.2
numpy==1.26.4
PyAudio==0.2.14
pydantic==2.7.1
PyQt5_sip==12.13.0
PySide2==5.15.2.1
PySide2==5.15.2.1
QDarkStyle==3.2.3
Requests==2.31.0
uvicorn==0.29.0
websocket_client==1.8.0
xpinyin==0.7.6
